backend/requirements.txt
# Core
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings

# Supabase
supabase

# Document Processing
pdfplumber
PyMuPDF
python-docx
Pillow
pytesseract
opencv-python
numpy

# NLP & ML
sentence-transformers
faiss-cpu
transformers
nltk
langdetect

# HTTP & Scraping
httpx
aiohttp
beautifulsoup4
playwright
lxml

# Workflow & Async
celery[redis]
redis

# Utilities
tenacity
python-jose[cryptography]
passlib[bcrypt]
